- sfml-tmxloader updated: now it hasn't anymore the code thrown in the project, but it was built as dynamic library and linked to the project - a funky version of the camera was written just to check the potentiality of the SFML library

+++ Game Engine 2D/GameObjectManager.cpp
@@ -49,7 +49,7 @@ Player* GameObjectManager::GetPlayer() const
 	return (Player* const)results->second;
 }
 
-int GameObjectManager::Count() const
+u_long GameObjectManager::Count() const
 {
 	return gameObjects.size();
 }
